Add basic driver implementation for the Grove-LCD RGB Backlight
The Grove-LCD RGB is made from two I2C client addresses, one for
addressing the LCD, another for the RGB backlight. This driver is
designed to handle both parts in one implementation, using the new API
"i2c_new_secondary_device". The driver targeting primarily the
HiKey960, the dev board used for the corresponding master's thesis.
This board has some issues on I2C which effects the LCD client's
address is not visible on the bus at the moment. To deal with this, the
driver uses the RGB client as main device and implements the LCD parts
as fail-safe as possible. Right now, they are rather conceptual due to
the hardware issues.