Add file operations for I2C driver
I2C driver, such as this one for the Grove-LCD RGB backlight, are a bit
difficult with file operations as its not (that common?) userfriendly
to restore private instance data within the file operation
implementations.
The trick is using container_of in the "open" call to restore the saved
private data via its "cdev" entry.