Change I2C device tree representation
By changing the device tree representation of the Grove's I2C slave
addresses to a single device with two regs, and exchange the primary
used one in the driver implementation, the driver became capable to
handle both parts just as intended. The most important part is the right
device tree representation, but due to some HiKey issues?, it is needed
to use the LCD as main device.