Add basic ioctl implementation
This commit adds the needed header definition for as save as possible
"ioctl"s. Use this headers also in userspace, e.g. as symlinks.
Currently, the ioctls are not fully implemented, the LCD write
operations are not ready yet, but the messaging between user and kernel
space is setted up and working safely.