Add ioctl implementation for actual display writes
The display has still some issues on Linux, but all at all, it is
working. The implemented behavior for the LCD writes is made from the
view of a correctly working Grove LCD. May we find the electrical issues
sometimes, but by taking a look at the oscilloscope, it is proofen, the
are correct. If not, there are pictures from the oscilloscope's output
on disk.
For the actual implementation, the 'struct string_t' was changed to hold
a position value instead of the string size (was not used in kernel) to
provide the same functionality as the Zircon driver variants.