Cleanup, style fixes and some minor fixes
This change take off the support for the "read()" and the "write()"
system calls, as the are not that meaningful for a composed
device/driver the actual one is.
Additionally, it fixes some code style issues that came from the Linux
cleanup scripts.
At last, it fixes some minor issues as forgotten error handling in
special situations, adds better comments at some locations and
alternative, newer APIs.